requirements/base.txt
fastapi>=0.115
uvicorn>=0.34
sqlalchemy>=2.0,<3.0
asyncpg>=0.30
colorlog>=6.7.0
psycopg2-binary>=2.9.3
python-dotenv>=1.0
pydantic-settings>=2.0
pydantic>=2.0,<3.0
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
email-validator>=2.0
cryptography>=44.0
alembic>=1.15
sqlalchemy-utils>=0.41
anyio>=4.0
httpx>=0.28
python-multipart>=0.0
